Point Makefile to newer driver
In theory, this may solve the de-sync issue with the new Series X controllers. On Windows 10, I discovered that the installed drivers for the dongle were dated 7/11/2017, when the newest available were dated 7/18/2018. Windows was reporting the older driver as the latest because it has a higher version number (1.0.46.1 vs 1.0.45.5). Forcing the driver update corrected the de-sync behavior in Windows.